requirements.txt
fastapi
uvicorn
python-keycloak
pydantic
pydantic-settings